backend/requirements.txt
fastapi==0.92.0
uvicorn==0.20.0
pdfkit==1.0.0
python-jose==3.3.0
psycopg[binary]==3.2.9
sqlmodel==0.0.24
alembic==1.16.1
python-slugify==8.0.1
jinja2==3.1.4